add integrity checks to manual validity checks
